+++ src/z80Registers.ts
@@ -5,14 +5,11 @@ import { Settings } from './settings';
 
 
 /// The formatting (for VARIABLES) for each register is provided through a map.
-export var Z80RegisterVarFormat: Map<string, string>;
+export let Z80RegisterVarFormat: Map<string, string>;
 
 /// The formatting (for hovering) for each register is provided through a map.
-export var Z80RegisterHoverFormat: Map<string, string>;
+export let Z80RegisterHoverFormat: Map<string, string>;
 
-/// All values of the registers are provided in a map.
-/// Together with a function to retrieve the value from the data string.
-var regMap = new Map<string, {(data: string):number}>();
 
 
 /**
@@ -22,9 +19,14 @@ export class Z80Registers {
 
 	/**
 	 * Eg.
-	 * PC=6005 SP=6094 BC=0100 AF=cf8c HL=02df DE=0fc9 IX=663c IY=5c3a AF'=0044 BC'=050e HL'=2758 DE'=0047 I=3f R=5e  F=S---3P-- F'=-Z---P-- MEMPTR=0000 IM1 IFF-- VPS: 0
+	 * PC=80cf SP=83f3 AF=0208 BC=0300 HL=4000 DE=2000 IX=ffff IY=5c3a AF'=0044 BC'=00ff HL'=f3f3 DE'=0001 I=00 R=0b  F=----3--- F'=-Z---P-- MEMPTR=0000 IM0 IFF12 VPS: 0
 	 * A65E RETI
 	 */
+
+	 /// All values of the registers are provided in a map.
+	/// Together with a function to retrieve the value from the data string.
+	protected static regMap = new Map<string, {(data: string):number}>();
+
 
 	// F flag constants for bit comparison.
 	public static FLAG_S = 1 << 7;
@@ -34,41 +36,57 @@ export class Z80Registers {
 	public static FLAG_N = 1 << 1;
 	public static FLAG_C = 1 << 0;
 
+	// Indices for first time search.
+	protected static pcIndex = -1;
+	protected static spIndex = -1;
+	protected static afIndex = -1;
+	protected static bcIndex = -1;
+	protected static hlIndex = -1;
+	protected static deIndex = -1;
+	protected static ixIndex = -1;
+	protected static iyIndex = -1;
+	protected static af2Index = -1;
+	protected static bc2Index = -1;
+	protected static hl2Index = -1;
+	protected static de2Index = -1;
+	protected static iIndex = -1;
+	protected static rIndex = -1;
+
 	/**
 	 * Called during the launchRequest.
 	 */
 	public static init() {
-		regMap.set("AF", Z80Registers.parseAF);
-		regMap.set("BC", Z80Registers.parseBC);
-		regMap.set("DE", Z80Registers.parseDE);
-		regMap.set("HL", Z80Registers.parseHL);
-		regMap.set("IX", Z80Registers.parseIX);
-		regMap.set("IY", Z80Registers.parseIY);
-		regMap.set("SP", Z80Registers.parseSP);
-		regMap.set("PC", Z80Registers.parsePC);
+		Z80Registers.regMap.set("AF", Z80Registers.parseAF);
+		Z80Registers.regMap.set("BC", Z80Registers.parseBC);
+		Z80Registers.regMap.set("DE", Z80Registers.parseDE);
+		Z80Registers.regMap.set("HL", Z80Registers.parseHL);
+		Z80Registers.regMap.set("IX", Z80Registers.parseIX);
+		Z80Registers.regMap.set("IY", Z80Registers.parseIY);
+		Z80Registers.regMap.set("SP", Z80Registers.parseSP);
+		Z80Registers.regMap.set("PC", Z80Registers.parsePC);
 
-		regMap.set("AF'", Z80Registers.parseAF2);
-		regMap.set("BC'", Z80Registers.parseBC2);
-		regMap.set("DE'", Z80Registers.parseDE2);
-		regMap.set("HL'", Z80Registers.parseHL2);
+		Z80Registers.regMap.set("AF'", Z80Registers.parseAF2);
+		Z80Registers.regMap.set("BC'", Z80Registers.parseBC2);
+		Z80Registers.regMap.set("DE'", Z80Registers.parseDE2);
+		Z80Registers.regMap.set("HL'", Z80Registers.parseHL2);
 
-		regMap.set("A", Z80Registers.parseA);
-		regMap.set("F", Z80Registers.parseF);
-		regMap.set("B", Z80Registers.parseB);
-		regMap.set("C", Z80Registers.parseC);
-		regMap.set("D", Z80Registers.parseD);
-		regMap.set("E", Z80Registers.parseE);
-		regMap.set("H", Z80Registers.parseH);
-		regMap.set("L", Z80Registers.parseL);
-		regMap.set("I", Z80Registers.parseI);
-		regMap.set("R", Z80Registers.parseR);
-		regMap.set("A'", Z80Registers.parseA2);
-		regMap.set("F'", Z80Registers.parseF2);
+		Z80Registers.regMap.set("A", Z80Registers.parseA);
+		Z80Registers.regMap.set("F", Z80Registers.parseF);
+		Z80Registers.regMap.set("B", Z80Registers.parseB);
+		Z80Registers.regMap.set("C", Z80Registers.parseC);
+		Z80Registers.regMap.set("D", Z80Registers.parseD);
+		Z80Registers.regMap.set("E", Z80Registers.parseE);
+		Z80Registers.regMap.set("H", Z80Registers.parseH);
+		Z80Registers.regMap.set("L", Z80Registers.parseL);
+		Z80Registers.regMap.set("I", Z80Registers.parseI);
+		Z80Registers.regMap.set("R", Z80Registers.parseR);
+		Z80Registers.regMap.set("A'", Z80Registers.parseA2);
+		Z80Registers.regMap.set("F'", Z80Registers.parseF2);
 
-		regMap.set("IXL", Z80Registers.parseIXL);
-		regMap.set("IXH", Z80Registers.parseIXH);
-		regMap.set("IYL", Z80Registers.parseIYL);
-		regMap.set("IYH", Z80Registers.parseIYH);
+		Z80Registers.regMap.set("IXL", Z80Registers.parseIXL);
+		Z80Registers.regMap.set("IXH", Z80Registers.parseIXH);
+		Z80Registers.regMap.set("IYL", Z80Registers.parseIYL);
+		Z80Registers.regMap.set("IYH", Z80Registers.parseIYH);
 
 		Z80RegisterVarFormat = Z80Registers.createFormattingMap(Settings.launch.formatting.registerVar);
 		Z80RegisterHoverFormat = Z80Registers.createFormattingMap(Settings.launch.formatting.registerHover);
@@ -83,18 +101,18 @@ export class Z80Registers {
 		const formattingMap = new Map();
 
 		// Read all formatting settings
-		for(var i=0; i<settingsMap.length; i+=2) {
-			var regRegex = new RegExp('^' + settingsMap[i] + '$');
-			var regFormat = settingsMap[i+1];
+		for(let i=0; i<settingsMap.length; i+=2) {
+			let regRegex = new RegExp('^' + settingsMap[i] + '$');
+			let regFormat = settingsMap[i+1];
 			// check for which registers the format should be used
-			for(let [key,] of regMap) {
+			for(let [key,] of Z80Registers.regMap) {
 				// get format
 				const format = formattingMap.get(key);
 				if(format != undefined)
 					continue;	// has already a format string
 				// now check if register is met
-				var keyWo = key;
-				var rLen = keyWo.length;
+				let keyWo = key;
+				let rLen = keyWo.length;
 				if(keyWo[rLen-1] == '\'')
 					keyWo = keyWo.substr(0, rLen-1);	// Remove the "'" in the register name
 				const match = regRegex.exec(keyWo);
@@ -106,7 +124,7 @@ export class Z80Registers {
 		}
 
 		// All unset registers get a default formatting
-		for(let [reg,] of regMap) {
+		for(let [reg,] of Z80Registers.regMap) {
 			// get format
 			const format = formattingMap.get(reg);
 			if(format != undefined)
@@ -138,74 +156,142 @@ export class Z80Registers {
 	 * @returns The value.
 	 */
 	public static parsePC(data: string): number {
-		const res = parseInt(data.substr(3,4),16);
+		if(Z80Registers.pcIndex < 0) {
+			Z80Registers.pcIndex = data.indexOf('PC=');
+			assert(Z80Registers.pcIndex >= 0);
+			Z80Registers.pcIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.pcIndex,4),16);
 		return res;
 	}
 
 	public static parseSP(data: string): number {
-		const res = parseInt(data.substr(11,4),16);
+		if(Z80Registers.spIndex < 0) {
+			Z80Registers.spIndex = data.indexOf('SP=');
+			assert(Z80Registers.spIndex >= 0);
+			Z80Registers.spIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.spIndex,4),16);
 		return res;
 	}
 
 	public static parseAF(data: string): number {
-		// AF
-		const res = parseInt(data.substr(27,4),16);
+		if(Z80Registers.afIndex < 0) {
+			Z80Registers.afIndex = data.indexOf('AF=');
+			assert(Z80Registers.afIndex >= 0);
+			Z80Registers.afIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.afIndex,4),16);
 		return res;
 	}
 
 	public static parseBC(data: string): number {
-		const res = parseInt(data.substr(19,4),16);
+		if(Z80Registers.bcIndex < 0) {
+			Z80Registers.bcIndex = data.indexOf('BC=');
+			assert(Z80Registers.bcIndex >= 0);
+			Z80Registers.bcIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.bcIndex,4),16);
 		return res;
 	}
 
 	public static parseHL(data: string): number {
-		const res = parseInt(data.substr(35,4),16);
+		if(Z80Registers.hlIndex < 0) {
+			Z80Registers.hlIndex = data.indexOf('HL=');
+			assert(Z80Registers.hlIndex >= 0);
+			Z80Registers.hlIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.hlIndex,4),16);
 		return res;
 	}
 
 	public static parseDE(data: string): number {
-		const res = parseInt(data.substr(43,4),16);
+		if(Z80Registers.deIndex < 0) {
+			Z80Registers.deIndex = data.indexOf('DE=');
+			assert(Z80Registers.deIndex >= 0);
+			Z80Registers.deIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.deIndex,4),16);
 		return res;
 	}
 
 	public static parseIX(data: string): number {
-		const res = parseInt(data.substr(51,4),16);
+		if(Z80Registers.ixIndex < 0) {
+			Z80Registers.ixIndex = data.indexOf('IX=');
+			assert(Z80Registers.ixIndex >= 0);
+			Z80Registers.ixIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.ixIndex,4),16);
 		return res;
 	}
 
 	public static parseIY(data: string): number {
-		const res = parseInt(data.substr(59,4),16);
+		if(Z80Registers.iyIndex < 0) {
+			Z80Registers.iyIndex = data.indexOf('IY=');
+			assert(Z80Registers.iyIndex >= 0);
+			Z80Registers.iyIndex += 3;
+		}
+		const res = parseInt(data.substr(Z80Registers.iyIndex,4),16);
 		return res;
 	}
 
 	public static parseAF2(data: string): number {
-		// AF'
-		const res = parseInt(data.substr(68,4),16);
+		if(Z80Registers.af2Index < 0) {
+			Z80Registers.af2Index = data.indexOf("AF'=");
+			assert(Z80Registers.af2Index >= 0);
+			Z80Registers.af2Index += 4;
+		}
+		const res = parseInt(data.substr(Z80Registers.af2Index,4),16);
 		return res;
 	}
 
 	public static parseBC2(data: string): number {
-		const res = parseInt(data.substr(77,4),16);
+		if(Z80Registers.bc2Index < 0) {
+			Z80Registers.bc2Index = data.indexOf("BC'=");
+			assert(Z80Registers.bc2Index >= 0);
+			Z80Registers.bc2Index += 4;
+		}
+		const res = parseInt(data.substr(Z80Registers.bc2Index,4),16);
 		return res;
 	}
 
 	public static parseHL2(data: string): number {
-		const res = parseInt(data.substr(86,4),16);
+		if(Z80Registers.hl2Index < 0) {
+			Z80Registers.hl2Index = data.indexOf("HL'=");
+			assert(Z80Registers.hl2Index >= 0);
+			Z80Registers.hl2Index += 4;
+		}
+		const res = parseInt(data.substr(Z80Registers.hl2Index,4),16);
 		return res;
 	}
 
 	public static parseDE2(data: string): number {
-		const res = parseInt(data.substr(95,4),16);
+		if(Z80Registers.de2Index < 0) {
+			Z80Registers.de2Index = data.indexOf("DE'=");
+			assert(Z80Registers.de2Index >= 0);
+			Z80Registers.de2Index += 4;
+		}
+		const res = parseInt(data.substr(Z80Registers.de2Index,4),16);
 		return res;
 	}
 
 	public static parseI(data: string): number {
-		const res = parseInt(data.substr(102,2),16);
+		if(Z80Registers.iIndex < 0) {
+			Z80Registers.iIndex = data.indexOf('I=');
+			assert(Z80Registers.iIndex >= 0);
+			Z80Registers.iIndex += 2;
+		}
+		const res = parseInt(data.substr(Z80Registers.iIndex,2),16);
 		return res;
 	}
 
 	public static parseR(data: string): number {
-		const res = parseInt(data.substr(107,2),16);
+		if(Z80Registers.rIndex < 0) {
+			Z80Registers.rIndex = data.indexOf('R=');
+			assert(Z80Registers.rIndex >= 0);
+			Z80Registers.rIndex += 2;
+		}
+		const res = parseInt(data.substr(Z80Registers.rIndex,2),16);
 		return res;
 	}
 
@@ -295,7 +381,7 @@ export class Z80Registers {
 		}
 		*/
 		const regUpper = reg.toUpperCase();
-		return regMap.get(regUpper) != undefined;
+		return Z80Registers.regMap.get(regUpper) != undefined;
 	}
 
 
@@ -318,7 +404,7 @@ export class Z80Registers {
 		const value = Z80Registers.getRegValueByName(reg, data);
 
 		// do the formatting
-		var rLen = reg.length;
+		let rLen = reg.length;
 		if(reg[rLen-1] == '\'') --rLen;	// Don't count the "'" in the register name
 
 		Utility.numberFormatted(reg, value, rLen, format, undefined, handler);
@@ -326,7 +412,7 @@ export class Z80Registers {
 
 
 	/**
-	 * Returns the 'variable' formatted register value.
+	 * Returns the 'letiable' formatted register value.
 	 * @param reg The name of the register, e.g. "A" or "BC"
 	 * @param data The data string returned by zesarux.
 	 * @param handler A function that is called with the formatted string as argument.
@@ -357,10 +443,9 @@ export class Z80Registers {
 	 * @returns The value of the register.
 	 */
 	public static getRegValueByName(regName: string, regsString:string): number {
-		//var handler = (data) => {};
-		var handler = regMap.get(regName.toUpperCase()) || (data => 0);
+		let handler = Z80Registers.regMap.get(regName.toUpperCase()) || (data => 0);
 		assert(handler != undefined, 'Register ' + regName + ' does not exist.');
-		var value = handler(regsString);
+		let value = handler(regsString);
 		return value;
 	}
 
@@ -373,10 +458,10 @@ export class Z80Registers {
 	 * @returns An array of strings with register names that match. If no matching register is found returns an empty array.
 	 */
 	public static getRegistersEqualTo(value: number, regsString: string): Array<string> {
-		var resRegs:Array<string> = [];
+		let resRegs: Array<string> = [];
 		if(regsString && regsString.length > 0) {
 			const regs = [ "HL", "DE", "IX", "IY", "SP", "BC", "HL'", "DE'", "BC'" ];
-			resRegs = regs.filter(reg => value == this.getRegValueByName(reg, regsString));
+			resRegs = regs.filter(reg => value == Z80Registers.getRegValueByName(reg, regsString));
 		}
 		return resRegs;
 	}
@@ -393,16 +478,16 @@ export class Z80Registers {
 		cc = (cc >> 1) & 0x03;
 		switch(cc) {
 			case 0:	// NZ, Z
-				condTest = ((flags & this.FLAG_Z) != 0);
+				condTest = ((flags & Z80Registers.FLAG_Z) != 0);
 				break;
 			case 1:	// NC, C
-				condTest = ((flags & this.FLAG_C) != 0);
+				condTest = ((flags & Z80Registers.FLAG_C) != 0);
 				break;
 			case 2:	// PO, PE
-				condTest = ((flags & this.FLAG_PV) != 0);
+				condTest = ((flags & Z80Registers.FLAG_PV) != 0);
 				break;
 			case 3:	// P, M
-				condTest = ((flags & this.FLAG_S) != 0);
+				condTest = ((flags & Z80Registers.FLAG_S) != 0);
 				break;
 			default:
 				assert(false);	// Impossible.